Update CategoryReportDetailsUiState -  val sumThisMonth: Double = 0.0,

+++ app/src/main/java/com/chskela/monoapplication/presentation/screens/details/models/CategoryReportDetailsUiState.kt
@@ -6,7 +6,7 @@ import com.chskela.monoapplication.presentation.screens.details.components.Repor
 data class CategoryReportDetailsUiState(
     val currentTab: Int = 0,
     val currentCategory: Long = 0,
-    val sumThisMonth: Long = 0,
+    val sumThisMonth: Double = 0.0,
     val categoryName: String = "",
     val icon: String = "",
     val typeCategory: TypeCategory = TypeCategory.Expense,
